loanapp/requirements.txt
django>=5.2,<6.0
django-countries>=7.5
pillow>=10.0
django-crispy-forms>=2.0
python-dotenv>=1.0
psycopg2-binary>=2.9
django-cleanup>=8.0
pytest-django>=4.5
django-allauth>=0.61
django-two-factor-auth>=1.15
django-otp>=1.2
qrcode>=7.4
requests>=2.31
cryptography>=45.0
colorama>=0.4
defusedxml>=0.7
charset-normalizer>=3.4
certifi>=2025.7
whitenoise>=6.6
gunicorn>=21.2